Fix nonce too low issue
